@@ -5,26 +5,47 @@
     "build": "next build",
     "start": "next start",
     "test": "jest --watch",
-    "test:ci": "jest --ci"
+    "test:ci": "jest --ci",
+    "lint": "next lint"
   },
   "dependencies": {
     "@apollo/client": "^3.11.8",
     "@heroicons/react": "^2.1.5",
+    "@tailwindcss/forms": "^0.5.9",
+    "@types/bcrypt": "^5.0.2",
     "@vercel/postgres": "^0.10.0",
     "apollo-server-micro": "^3.13.0",
+    "autoprefixer": "^10.4.20",
+    "bcrypt": "^5.1.1",
     "clsx": "^2.1.1",
+    "css-loader": "^7.1.2",
+    "form-data": "^4.0.0",
     "graphql": "^16.9.0",
     "graphql-tag": "^2.12.6",
-    "next": "latest",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "server-only": "^0.0.1"
+    "next": "^15.0.0-rc.0",
+    "next-auth": "5.0.0-beta.20",
+    "postcss": "^8.4.45",
+    "postcss-loader": "^8.1.1",
+    "react": "^19.0.0-rc-94e652d5-20240912",
+    "react-dom": "^19.0.0-rc-94e652d5-20240912",
+    "sass": "^1.78.0",
+    "sass-loader": "^16.0.1",
+    "server-only": "^0.0.1",
+    "style-loader": "^4.0.0",
+    "tailwindcss": "^3.4.11",
+    "use-debounce": "^10.0.3",
+    "zod": "^3.23.8",
+    "@types/react": "npm:types-react@rc",
+    "@types/react-dom": "npm:types-react-dom@rc"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/react": "^14.1.2",
     "@types/jest": "^29.5.11",
+    "@types/node": "^22.5.4",
     "@types/react": "18.2.45",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.11",
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
     "typescript": "5.3.3"
