@@ -1,6 +1,6 @@
 import Link from 'next/link';
-import NavLinks from '../dashboard/nav-links';
-import AcmeLogo from '../acme-logo';
+import NavLinks from './nav-links';
+//import AcmeLogo from '../acme-logo';
 import { PowerIcon } from '@heroicons/react/24/outline';
 
 export default function SideNav() {
@@ -10,9 +10,9 @@ export default function SideNav() {
         className="mb-2 flex h-20 items-end justify-start rounded-md bg-blue-600 p-4 md:h-40"
         href="/"
       >
-        <div className="w-32 text-white md:w-40">
+        {/* <div className="w-32 text-white md:w-40">
           <AcmeLogo />
-        </div>
+        </div> */}
       </Link>
       <div className="flex grow flex-row justify-between space-x-2 md:flex-col md:space-x-0 md:space-y-2">
         <NavLinks />
