@@ -1,23 +1,51 @@
+// In the Search component file (e.g., `search.tsx`)
+"use client";
 import { MagnifyingGlassIcon } from '@heroicons/react/24/outline';
+import { usePathname, useRouter, useSearchParams } from 'next/navigation';
+import { useDebouncedCallback } from 'use-debounce';
 
-interface SearchProps {
+export default function Search({
+  placeholder,
+  onChange,
+}: {
   placeholder: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-}
+  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void; // Add the onChange prop
+}) {
+  const searchParams = useSearchParams();
+  const { replace } = useRouter();
+  const pathname = usePathname();
 
-export default function Search({ placeholder, onChange }: SearchProps) {
+  const handleSearch = useDebouncedCallback((term: string) => {
+    console.log(`Searching... ${term}`);
+
+    const params = new URLSearchParams(searchParams);
+
+    params.set('page', '1');
+
+    if (term) {
+      params.set('query', term);
+    } else {
+      params.delete('query');
+    }
+    replace(`${pathname}?${params.toString()}`);
+  }, 300);
+
   return (
     <div className="relative flex flex-1 flex-shrink-0">
       <label htmlFor="search" className="sr-only">
         Search
       </label>
       <input
-        id="search"
-        type="text"
+        id="search" // Added id to match label htmlFor
         className="peer block w-full rounded-md border border-gray-200 py-[9px] pl-10 text-sm outline-2 placeholder:text-gray-500"
         placeholder={placeholder}
-        onChange={onChange}
+        onChange={(e) => {
+          handleSearch(e.target.value);
+          if (onChange) onChange(e); // Call onChange if it's provided
+        }}
+        defaultValue={searchParams.get('query')?.toString()}
       />
+      <MagnifyingGlassIcon className="absolute left-3 top-1/2 h-[18px] w-[18px] -translate-y-1/2 text-gray-500 peer-focus:text-gray-900" />
     </div>
   );
 }
