@@ -1,5 +1,6 @@
 import { PencilIcon, PlusIcon, TrashIcon } from '@heroicons/react/24/outline';
 import Link from 'next/link';
+import { deleteTask } from '../../lib/actions';
 
 export function CreateTask() {
   return (
@@ -16,7 +17,7 @@ export function CreateTask() {
 export function UpdateTask({ id }: { id: string }) {
   return (
     <Link
-      href="/dashboard/tasks"
+      href={`/dashboard/tasks/${id}/edit`}
       className="rounded-md border p-2 hover:bg-gray-100"
     >
       <PencilIcon className="w-5" />
@@ -25,12 +26,14 @@ export function UpdateTask({ id }: { id: string }) {
 }
 
 export function DeleteTask({ id }: { id: string }) {
+  const deleteTaskWithId = deleteTask.bind(null, id);
+
   return (
-    <>
-      <button className="rounded-md border p-2 hover:bg-gray-100">
+    <form action={deleteTaskWithId}>
+      <button type="submit" className="rounded-md border p-2 hover:bg-gray-100">
         <span className="sr-only">Delete</span>
         <TrashIcon className="w-5" />
       </button>
-    </>
+    </form>
   );
 }
